# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QaCheckOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "id",
          "passed",
          "reasons"
        ],
        "title": "QaCheckOut",
        "type": "object"
      },
      "QaResultOut": {
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/QaCheckOut"
            },
            "title": "Checks",
            "type": "array"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "video_id": {
            "format": "uuid",
            "title": "Video Id",
            "type": "string"
          }
        },
        "required": [
          "video_id",
          "passed",
          "checks"
        ],
        "title": "QaResultOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CreatorIQX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/videos/{video_id}/qa": {
      "get": {
        "operationId": "get_qa_api_v1_videos__video_id__qa_get",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read the current QA checklist outcome",
        "tags": [
          "qa"
        ]
      }
    },
    "/api/v1/videos/{video_id}/qa/run": {
      "post": {
        "operationId": "run_qa_api_v1_videos__video_id__qa_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Video Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QaResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run the pre-publish QA checklist",
        "tags": [
          "qa"
        ]
      }
    },
...
  }
}